Disable UnsafeMemory on big endian machines
The current UnsafeMemory code assumes the machine is little endian, and will
not work correctly on big endian CPUs.  For now, we will disable UnsafeMemory on
big endian machines.  This will make the code significantly slower on big endian.
In the future someone should add the necessary flip bytes calls to make this
work efficiently on big endian machines.